rewrote the backend cursor handling
WT Cursor must be re-used. During the generation of a list of items
from a cursor, a cursor is retrieved and kept "open" for the current
generator until that generator is fully consumed. Problem is the generator
is not always fully consumed because of `gremlin.limit` without a way to
close the iterator/generator cursors will leak each time a "cursor generator"
is started but "left behind" before full consumption. Previously I was passing
`CursorManager.reset` method in the result, so that the gremlin step could
pass or close the reset method when it found that it doesn't need it anymore.
the problem is that this breaks the gremlin API of steps.

Steps must be very simple to implement `step(graphdb, iterator)`.

The new implementation use the iterator protocol, so that generators bound
to a cursor have an extra method `close` which will reset the cursor and
avoid the leak. This way, there is no extra argument to pass along, but the
gremlin steps must be decorated to return a closeable generator. The idea
is to automatically generate `CloseableIterator` for each steps.

If this works this will keep the implementation of new steps simple while
avoiding any cursor leak.